requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
dj-database-url==0.5.0
Django==3.2.6
django-cors-headers==3.7.0
django-editorjs==0.2.1
django-editorjs-fields==0.2.2
django-heroku==0.0.0
djangorestframework==3.12.4
gunicorn==20.1.0
idna==3.2
Markdown==3.3.4
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
whitenoise==5.3.0
